set up components and loadQueue function in Mine
