+++ client/src/utils/API.js
@@ -2,9 +2,9 @@ import axios from "axios";
 
 export default {
   //thinking of this as createQueue() to match the action type
-  getData: function (query) {
+  getQueue: function (topics) {
     //for now, just hardcoded for it to return 50 results
-    let queryURL = `https://microsoft-azure-bing-news-search-v1.p.rapidapi.com/search?q=${query}&count=50&sortby=dat`;
+    let queryURL = `https://microsoft-azure-bing-news-search-v1.p.rapidapi.com/search?q=${topics}&count=50&sortby=dat`;
 
     const config = {
       method: "get",
@@ -19,6 +19,4 @@ export default {
       return res;
     });
   },
-
-  
 };
